refactor: updating the utils and imports
