+++ components/selectedEthereumAccount.tsx
@@ -1,6 +1,7 @@
 import { useConnectEthereumWallet } from "@/hooks/useConnectEthereumWallet";
 import { useSwitchEthereumNetwork } from "@/hooks/useSwitchEthereumNetwork";
-import { cn, trimAccount } from "@/lib/utils";
+import { cn } from "@/utils/cn";
+import { trimAccount } from "@/utils/formatting";
 import { ethereumAccountAtom } from "@/store/ethereum";
 import { useAtomValue } from "jotai";
 import { Button } from "./ui/button";
